# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "laravel",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "alaarezk322@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/blogs": {
            "get": {
                "tags": [
                    "blogs"
                ],
                "summary": "Get list of blogs",
                "operationId": "9819e5802278e060898795fe3c919599",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "blogs"
                ],
                "summary": "Create a blog",
                "operationId": "514d226035b3d74b6ccf9837b3d1ed36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    }
                }
            }
        },
        "/api/blogs/{id}": {
            "get": {
                "tags": [
                    "blogs"
                ],
                "summary": "Get a specific blog",
                "operationId": "edf2058da1858ee867ebc4789b3195f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "blogs"
                ],
                "summary": "Update a specific blog",
                "operationId": "6a99affa8f7d94cd429d58eb73cf0e88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Blog updated successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "blogs"
                ],
                "summary": "Delete a specific blog",
                "operationId": "9fb0bb3fc65e987a03d2b5ecdda4489d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the blog",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blog deleted successfully"
                    },
                    "404": {
                        "description": "Blog not found"
                    }
                }
            }
        },
...
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "description": "Fetch a single product by its ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "description": "Update a product with the provided data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "subtitle",
                                    "launch_date",
                                    "company_name",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "subtitle": {
                                        "type": "string"
                                    },
                                    "launch_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "video": {
                                        "type": "string",
                                        "format": "uri"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "description": "Delete a product by its ID",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
...
}